add Registration with Authorization

+++ my_app_run/src/main/java/com/example/juridov/my_app_run/service/RecordService.java
@@ -36,6 +36,9 @@ public class RecordService {
     }
 
     public void delete(Long id) {
+        if (recordRepository.findRecordById(id) == null){
+            return;
+        }
         recordRepository.delete(recordRepository.findRecordById(id));
     }
 }
